+++ src/compiler/docs/readme/output-docs.ts
@@ -47,12 +47,23 @@ export const generateReadme = async (
         const readmeOutputPath = join(readmeOutput.dir, relativeReadmePath);
 
         const currentReadmeContent =
-          normalizePath(readmeOutput.dir) !== normalizePath(config.srcDir)
-            ? // The user set a custom `.dir` property, which is where we're going
-              // to write the updated README. We need to read the non-automatically
-              // generated content from that file and preserve that.
-              await getUserReadmeContent(compilerCtx, readmeOutputPath)
-            : userContent;
+          readmeOutput.overwriteExisting === true
+            ? // Overwrite explicitly requested: always use the provided user content.
+              userContent
+            : normalizePath(readmeOutput.dir) !== normalizePath(config.srcDir)
+              ? (readmeOutput.overwriteExisting === 'if-missing' &&
+                  // Validate a file exists at the output path
+                  (await compilerCtx.fs.access(readmeOutputPath))) ||
+                // False and undefined case: follow the changes made in #5648
+                (readmeOutput.overwriteExisting ?? false) === false
+                ? // Existing file found: The user set a custom `.dir` property, which is
+                  // where we're going to write the updated README. We need to read the
+                  // non-automatically generated content from that file and preserve that.
+                  await getUserReadmeContent(compilerCtx, readmeOutputPath)
+                : // No existing file found: use the provided user content.
+                  userContent
+              : // Default case: writing to srcDir, so use the provided user content.
+                userContent;
 
         const readmeContent = generateMarkdown(currentReadmeContent, docsData, cmps, readmeOutput, config);